requirements.txt
fastapi
uvicorn
jinja2
python-multipart
PyPDF2
reportlab
google-api-python-client
google-auth
google-auth-oauthlib
google-auth-httplib2
requests
python-dotenv
